telegram-api/pyproject.toml
[project]
name = "telegram-api"
version = "0.1.0"
description = ""
authors = [
    {name = "bani2016up",email = "bani2016up@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic (>=2.11.10,<3.0.0)",
    "aio-pika (>=9.5.7,<10.0.0)"
]
[tool.poetry]
packages = [{include = "telegram_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
